move `representability` out of `rustc_middle`
